+++ pkg/utils/case.go
@@ -1,0 +1,23 @@
+package utils
+
+import "strings"
+
+// toSnakeCase converts PascalCase to snake_case
+func ToSnakeCase(s string) string {
+	var result strings.Builder
+	for i, r := range s {
+		if i > 0 && r >= 'A' && r <= 'Z' {
+			result.WriteRune('_')
+		}
+		result.WriteRune(r)
+	}
+	return strings.ToLower(result.String())
+}
+
+// toTitleCase converts a string to title case (simple implementation)
+func ToTitleCase(s string) string {
+	if len(s) == 0 {
+		return s
+	}
+	return strings.ToUpper(s[:1]) + s[1:]
+}